First commit
Contains a bare bones announcement system, a button for debugging use, and some tabs that we can start filling in with UI. TTS is working, but does NOT initialize on emulators it seems. Debug with a physical device.
Comments have instructions and information, change announceText via announceText.setValue("value"). NOTE: Changing that variable WILL cause it to announce the new value with text to speech right now. I have yet to implement the controls needed to prevent spam, and it WILL queue up these changes for the time being. If tested with OBD/GPS integration, change the queue mode in speakText() to QUEUE_FLUSH. Ask me if you need help.